+++ src/pages/Contact/Contact.js
@@ -1,15 +1,10 @@
 import React from 'react';
+import SampleContent from 'components/SampleContent';
 
 const ContactPage = () => (
   <div>
     <h2>Contact Us</h2>
-    <p>
-      Spare ribs short ribs ham sirloin. Sausage andouille filet mignon frankfurter spare ribs sirloin buffalo
-      t-bone hamburger drumstick strip steak brisket turducken jowl leberkas. Kielbasa meatball bacon tongue pork
-      loin. Ham tri-tip pancetta, fatback ball tip andouille jowl porchetta ground round. Prosciutto bacon ham hock
-      bresaola meatball chicken sirloin pork tongue burgdoggen pork chop. Rump kielbasa jowl chicken, filet mignon
-      fatback buffalo meatball spare ribs beef tri-tip.
-    </p>
+		<SampleContent paragraphLength={5} />
   </div>
 );