+++ src/reducers/index.js
@@ -1,6 +1,8 @@
 import { combineReducers } from 'redux';
 import notifications from './notificationReducer';
+import modal from './modalReducer';
 
 export default combineReducers({
-	notifications
+	notifications,
+	modal
 });