+++ src/reducers/modalReducer.js
@@ -1,0 +1,16 @@
+import {
+	SHOW_MODAL,
+	HIDE_MODAL
+} from '../constants/modalConstants';
+import initialState from '../state/modalState';
+
+export default (state = initialState , action) => {
+	switch (action.type) {
+		case SHOW_MODAL:
+			return action.payload;
+		case HIDE_MODAL:
+			return initialState;
+		default:
+			return state;
+	}
+};