+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # React Redux Jumpstart
 
+#### v.0.6.3
+* *Feature*: Added SampleContent component
+
+#### v.0.6.2
+* *Update*: Added Modal functionality to Redux
+
+#### v.0.6.1
+* *Feature*: Added random content util
+
 #### v.0.6.0
 * *Feature*: Added Modal component
 * *Feature*: Added Dismiss component